requirements.txt
appdirs==1.4.3
blinker==1.4
click==6.7
dominate==2.3.1
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
six==1.10.0
SQLAlchemy==1.1.9
SQLAlchemy-JSONField==0.6.3
ujson==1.35
visitor==0.1.3
Werkzeug==0.12.1
WTForms==2.1
